feat: added functionality to create and view apartments and rooms (#main)
This commit adds new features to the application that enable users to create and view apartments and rooms.
It includes the creation of a new form to add apartments and rooms, which is integrated with the Supabase backend to store the data.
The commit also modifies the apartment and room viewing pages to fetch data from the backend and display it.
Furthermore, the commit includes the addition of the Supabase client and the configuration for the Next.js image component to accept images from the Supabase storage.
Lastly, a new 'types' file was created to define the types used in the application, and the .env file was added to the .gitignore to avoid exposing sensitive data.